requirements.txt
aiodns==3.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
backports.zstd==1.0.0
base58==2.1.1
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chutes==0.3.60
click==8.3.0
cryptography==46.0.3
cytoolz==1.1.0
distro==1.9.0
ecdsa==0.19.1
eth-hash==0.7.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.1
fastapi==0.121.1
fickling==0.1.3
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
graval==0.2.6
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
jiter==0.12.0
loguru==0.7.2
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.0
netifaces==0.11.0
openai==2.7.2
orjson==3.11.4
packaging==25.0
prometheus_client==0.21.0
propcache==0.4.1
psutil==7.1.3
py-bip39-bindings==0.3.0
py-ed25519-zebra-bindings==1.3.0
py-sr25519-bindings==0.2.2
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
python-dotenv==1.2.1
pyudev==0.24.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scalecodec==1.2.12
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
substrate-interface==1.7.11
toolz==1.1.0
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websocket-client==1.9.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
